+++ src/node/variable.cpp
@@ -8,17 +8,17 @@ std::string Variable::getDefaultTag() {
     return "variable_";
 }
 
-Variable::Variable(const float init, std::string n_tag) {
+Variable::Variable(const double init, std::string n_tag) {
     createTag(n_tag);
     initial_value = Tensor(init);
 }
 
-Variable::Variable(const std::vector<float> data, const std::vector<int> shape, std::string n_tag) {
+Variable::Variable(const std::vector<double> data, const std::vector<unsigned int> shape, std::string n_tag) {
     createTag(n_tag);
     initial_value = Tensor(data, shape);
 }
 
-Variable::Variable(const std::initializer_list<float>& data, const std::initializer_list<int>& shape, std::string n_tag) {
+Variable::Variable(const std::initializer_list<double>& data, const std::initializer_list<unsigned int>& shape, std::string n_tag) {
     createTag(n_tag);
     initial_value = Tensor(data, shape);
 }
